refactor: rename button.styles.ts to have same naming
